Fixed the improved fast deployment support.
What happened there was: the latest Xamarin.Android product completely broke
whatever I have implemented in the past. It used to be that we can simply
specify `xbuild /p:AndroidFastDeploymentType=assemblies:dexes`, but now
they have constant `<_InstantRunEnabled>false</_InstantRunEnabled>` which
completely disables this awesome feature.

Since the property is defined in the targets and ignores the possibility
that it can be overriden (anything that starts with _ should not be
externally configurable, it's just that whoever made this change is evil),
specifying the property in the .props file doesn't work.
Therefore, this changeset specifies it at .targets.